AUT-575, AUT-576, AUT-577, AUT-582 (#22)
* feat: Added INCA MFA tools

Transferred from personal repository: .py file with relevant functions, an example case with simplified notebook, tests and test data, a short description in the docs, addition to the requirements

* fix: adjusted names to import functions

* fix: adjusted using black and quick fix for line length issues

Added "# noqa E501" to lines that were to long where I could not just break them (the strings that are necessary for setting up the MATLAB file) because breaking them caused issues with the resulting script

* fix: forgot one file

* fix: Adjusted according to comments

"Can you add the argument and return sections to the docstrings?"

* fix: adjustment of requirements.txt according to comments

"I think pytfa installs numpy and pandas. Since there is going to be more requirements in this repo I propose to specify the version of pandas and numpy if you think that it is necessary. Cobra sometimes has incompatibilities with wrong version of numpy or pandas."

* fix: corrected typos

Typos that were pointed in PR comments have been adjusted

* feat: biomass function is now an editable parameter

* fix: moved comments to not mess with docstrings

* feat: the unbalanced metabolites are now a parameter

* fix: changes script for runner

* fix: changed because of addition of biomass function

* feat: added biomass function as a parameter

* fix: adapted and generated test data

* fix: added error message if there is a duplicate reaction

* Update INCA_script_generator.py

* feat: changed test data generation and added the new obj

* tests: added new tests for most of the custom functions

* fix: black caused an issue with the data import

* tests: small adjustments and comments

* fix: test if it still works

* fix: biomass function can be imported from .csv file

* tests: adjusted tests file and data (biomass from .csv)

* fix: adjusted example notebook

* fix: oops, whitespace

* feat: moved everything to a class

Also moved the script to a separate folder

* fix: updated test script

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* Update test_INCA_script_generator.py

* fix: updated example notebook